server/requirements.txt
openenv-core>=0.2.0
pandas>=2.0
numpy>=1.24
fastapi>=0.100
uvicorn>=0.20
pydantic>=2.0
